Created NEAT Classes and Reproduction algorithms
These are for the original NEAT algorithm, and not for the CoDeepNEAT algorithm.  Just wanted to get a starting point established to start developing CoDeepNEAT from.

The NEAT_Testing.py file is a crude test of the crossover algorithm, using the example in Figure 4 (pg. 11) from the NEAT paper.  Seems to be working.